[feature] MCP connection popover 加上 endpoint copy 按鈕
把 SettingsApp.tsx 裡的 CopyButton 抽成共用元件，讓 McpStatusChip 的
連線 popover 也能複製 MCP endpoint URL，不用再手動選取文字。
