+++ src/settings/SettingsApp.tsx
@@ -6,7 +6,7 @@ import { appLogDir, join } from "@tauri-apps/api/path";
 import { revealItemInDir } from "@tauri-apps/plugin-opener";
 import { toast } from "sonner";
 import { log } from "../lib/log";
-import { Check, Copy, Download, Loader2, LogIn, LogOut, Monitor, Moon, PlugZap, Plus, ScrollText, Sun, Trash2 } from "lucide-react";
+import { Check, Download, Loader2, LogIn, LogOut, Monitor, Moon, PlugZap, Plus, ScrollText, Sun, Trash2 } from "lucide-react";
 import { useStore } from "../lib/store";
 import { LANGUAGE_OPTIONS, useI18n, type TranslationKey } from "../i18n";
 import { broadcastSettings } from "../lib/settingsSync";
@@ -32,6 +32,7 @@ import { ReleaseNotesDialog } from "../components/ReleaseNotesDialog";
 import { UsagePanel } from "./UsagePanel";
 import { STT_PROVIDERS, STT_BY_ID } from "../lib/transcription/providers";
 import { Button } from "@/components/ui/button";
+import { CopyButton } from "@/components/CopyButton";
 import { Toaster } from "@/components/ui/sonner";
 import { Input } from "@/components/ui/input";
 import { PasswordInput } from "@/components/ui/password-input";
@@ -1175,54 +1176,6 @@ function AccountSignInField({
         </div>
       )}
     </Field>
-  );
-}
-
-/** Copy-to-clipboard button with a 2s "copied" confirmation. */
-function CopyButton({
-  value,
-  label,
-  title,
-  className,
-  iconOnly,
-  disabled,
-}: Readonly<{
-  /** Text to copy, or a thunk evaluated at click time for computed values. */
-  value: string | (() => string);
-  label?: string;
-  title?: string;
-  className?: string;
-  iconOnly?: boolean;
-  disabled?: boolean;
-}>) {
-  const { t } = useI18n();
-  const [copied, setCopied] = useState(false);
-  const copy = () => {
-    navigator.clipboard.writeText(typeof value === "function" ? value() : value);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-  };
-  if (iconOnly) {
-    return (
-      <Button variant="ghost" size="icon" className={className} title={title} disabled={disabled} onClick={copy}>
-        {copied ? <Check className="size-3 text-emerald-500" /> : <Copy className="size-3" />}
-      </Button>
-    );
-  }
-  return (
-    <Button variant="outline" size="sm" className={className} title={title} disabled={disabled} onClick={copy}>
-      {copied ? (
-        <>
-          <Check className="size-3.5 text-emerald-500" />
-          <span>{t("settings.mcp.copied")}</span>
-        </>
-      ) : (
-        <>
-          <Copy className="size-3.5" />
-          <span>{label}</span>
-        </>
-      )}
-    </Button>
   );
 }
